refactor(migration): merge rsync-fix-encoding into rsync-migration
