+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["rsync-core", "rsync-fetcher", "rsync-fix-encoding", "rsync-gateway", "rsync-gc", "rsync-migration"]
+members = ["rsync-core", "rsync-fetcher", "rsync-gateway", "rsync-gc", "rsync-migration"]
 resolver = "2"
 
 [profile.dev.package.sqlx-macros]
