--- rsync-fix-encoding/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "rsync-fix-encoding"
-version = "0.1.0"
-edition = "2021"
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-bincode = "=2.0.0-rc.3"
-bincode_old = { package = "bincode", git = "https://github.com/bincode-org/bincode", rev = "v2.0.0-rc.2" }
-clap = { version = "4.1", features = ["derive"] }
-color-eyre = "0.6"
-eyre = "0.6"
-futures = "0.3"
-indicatif = "0.17"
-redis = { git = "https://github.com/PhotonQuantum/redis-rs", features = ["aio", "tokio-comp"] }
-rsync-core = { path = "../rsync-core", features = ["redis"] }
-tokio = { version = "1.25", features = ["full"] }
-tracing = "0.1"
-url = "2.3"
