+++ rsync-migration/Cargo.toml
@@ -5,6 +5,8 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+bincode = "=2.0.0-rc.3"
+bincode_old = { package = "bincode", git = "https://github.com/bincode-org/bincode", rev = "v2.0.0-rc.2" }
 clap = { version = "4.3", features = ["derive", "env"] }
 dotenvy = "0.15"
 eyre = "0.6"
