+++ rsync-migration/src/main.rs
@@ -4,9 +4,11 @@ use rsync_core::utils::{init_color_eyre, init_logger};
 
 use crate::redis_to_pg::redis_to_pg;
 use crate::schema_migration::schema_migration;
+use crate::upgrade_encoding::upgrade_encoding;
 
 mod redis_to_pg;
 mod schema_migration;
+mod upgrade_encoding;
 
 /// rsync-sjtug migration tool.
 #[derive(Parser)]
@@ -19,8 +21,10 @@ pub struct Args {
 pub enum Command {
     /// Apply database schema migrations.
     SchemaMigration(schema_migration::Args),
-    /// Migrate data from redis to postgresql (pre 0.4 -> 0.4).
+    /// Migrate data from redis to postgresql (0.3 -> 0.4).
     RedisToPg(redis_to_pg::Args),
+    /// Upgrade Redis index encoding (0.2 -> 0.3).
+    UpgradeEncoding(upgrade_encoding::Args),
 }
 
 #[tokio::main]
@@ -35,6 +39,7 @@ async fn main() -> eyre::Result<()> {
     match args.cmd {
         Command::SchemaMigration(args) => schema_migration(args).await?,
         Command::RedisToPg(args) => redis_to_pg(args).await?,
+        Command::UpgradeEncoding(args) => upgrade_encoding(args).await?,
     }
 
     Ok(())
